fix(core): stop uffsd crash on multi-extension --ext filters + close UTF-16 anti-pattern gaps
resolve_ext_ids zipped an unbounded 0_u16.. range against ext_names.iter();
RangeFrom<u16>::next() computes its next state eagerly, so once a drive's
ext_names table reached its legitimate u16::MAX ceiling, a query extension
absent from that drive drove the range one step past u16::MAX and panicked
with "attempt to add with overflow" - taking the whole daemon down. Switched
to .enumerate(), which cannot overflow. Also closes two from_utf16_lossy
anti-pattern gaps: $ATTRIBUTE_LIST/$UsnJrnl:$J name decoding now routes
through the crate's shared malformed-name-safe decoder (WI-4.1) instead of
a redundant lossy implementation, and the two genuinely non-filename Win32
sysinfo decodes (volume label, registry string) are marked AUDIT-OK.

Also documents the --agg <PRESET> top-N cap and its terms:FIELD,top=N
escape hatch in --agg --help, and splits args.rs's static help text into
args_help.rs to stay under the workspace's 800-LOC file policy.
